feat(playground): use new bootstrap and move todo to feature folder
