+++ playground/app/index.ts
@@ -1,19 +1,11 @@
 import * as angular from 'angular';
 import { provide } from 'ng-metadata/core';
-import { Title } from 'ng-metadata/platform';
 
 import { TabsModule } from './components/tabs/index'
 import { LifecycleHooksModule } from './components/lifecycle/index';
 import { ChangeDetectorModule } from './components/change-detector/index';
 import { TitleHandlerModule } from './components/title/index';
 
-import { TodoAppCmp } from './components/todo-app.component';
-import { TodoItemCmp } from './components/todo-item.component';
-import { AddTodoCmp } from './components/add-todo.component';
-
-import { RemainingTodosPipe } from './pipes/remainingTodos.pipe';
-
-import { TodoStore } from './stores/todoStore.service';
 
 import { ElementReadyDirective } from './directives/element-ready.directive';
 import { MyValidatorDirective } from './directives/my-validator.directive';
@@ -29,15 +21,6 @@ export const AppModule = angular.module( 'app', [
   ChangeDetectorModule,
   TitleHandlerModule
 ] )
-
-  // we need to register the service manually
-  .service( ...provide( Title ) )
-
-  .directive( ...provide( TodoAppCmp ) )
-  .directive( ...provide( AddTodoCmp ) )
-  .directive( ...provide( TodoItemCmp ) )
-  .filter( ...provide( RemainingTodosPipe ) )
-  .service( ...provide( TodoStore ) )
 
   .directive( ...provide( ElementReadyDirective ) )
