+++ playground/app/main.ts
@@ -1,8 +1,11 @@
 //main entry point
-import {bootstrap} from 'ng-metadata/platform';
+import { bootstrap } from 'ng-metadata/platform';
+import { Title } from 'ng-metadata/platform';
 import {enableProdMode} from 'ng-metadata/core';
-import {AppModule} from './index';
+
+import { AppComponent } from './app.component';
+import { AppModule } from './index';
 
 // enableProdMode();
 
-bootstrap( AppModule );
+bootstrap( AppComponent, [ Title, AppModule ] );
