+++ playground/app/todo/todo-item.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, Output, OnChanges, EventEmitter } from 'ng-metadata/core';
-import { TodoModel } from '../stores/todoStore.service';
+import { TodoModel } from './todo-store.service';
 
 @Component({
   selector: 'todo-item',
-  templateUrl: './app/components/todo-item.html',
+  templateUrl: './app/todo/todo-item.html',
   legacy: {
     transclude: true
   }
@@ -14,7 +14,7 @@ export class TodoItemCmp implements OnChanges{
   @Input('<') idx: number;
   // this is old way how to bindings are created, will be removed in 2.0
   // @Output() onDone: ( todo: {todo:TodoModel} )=>void;
-  @Output() onDone: EventEmitter<TodoModel>; // used as an interface 
+  @Output() onDone: EventEmitter<TodoModel>; // used as an interface
   @Output() emitTest = new EventEmitter<string>(); // use class instance
 
   ngOnInit(){
@@ -26,7 +26,7 @@ export class TodoItemCmp implements OnChanges{
     } )
 
   }
-  
+
   ngOnChanges(change){
     console.log('changes:', change);
     if(change.todo){
