+++ playground/index.html
@@ -70,9 +70,9 @@
           class="demo-content mdl-color--white mdl-shadow--4dp content mdl-color-text--grey-800 mdl-cell mdl-cell--8-col">
 
           <!--APP Component-->
-          <todo-app style="display: block" element-ready my-global-listener>
+          <my-app style="display: block" element-ready my-global-listener>
             <div style="text-align: center"><div class="mdl-spinner mdl-js-spinner is-active"></div></div>
-          </todo-app>
+          </my-app>
 
         </div>
       </div>
